+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Subsystems/DetectMarker.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Subsystem;
+package org.firstinspires.ftc.teamcode.Subsystems;
 
 import com.qualcomm.robotcore.hardware.HardwareMap;
