+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Subsystems/MarkerLocation.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Subsystem;
+package org.firstinspires.ftc.teamcode.Subsystems;
 
 /**
  * The Marker Location is {@link #SEARCHING} by default.
